feat(auth): real window.nostr interop via dart:js_interop
Splits the previous always-null stub into:
- nip07_types.dart: shared type definitions
- nip07_interop_stub.dart: non-web (isAvailable false)
- nip07_interop_web.dart: real dart:js_interop binding to window.nostr
- nip07_interop.dart: thin dispatcher selecting via conditional import

The web variant binds getPublicKey, signEvent, getRelays, plus nip04
and nip44 encrypt/decrypt. Nip07Service's public surface is unchanged
because it only imports nip07_interop.dart and only uses the names
re-exported there.
