+++ mobile/lib/services/nip07_interop_stub.dart
@@ -1,0 +1,7 @@
+// ABOUTME: Non-web stub for NIP-07 interop. Always reports unavailable.
+
+import 'package:openvine/services/nip07_types.dart';
+
+NostrExtension? get nostr => null;
+
+bool get isNip07Available => false;
